Initialize ob_type for class type objects early
When creating classes, the `ob_type` field of the class's `PyTypeObject`
is statically initialized to `NULL`, and then becomes dynamically
initialized to `&PyType_Type` when `PyType_Ready()` is called.  This is
too late.  If GC occurs during the initialization of the class before
`PyType_Ready()` is called, then the garbage collector will encounter
the `NULL` `ob_type` and segfault.

The Python docs say this field should be dynamically initialized to
`&PyType_Type` before doing any more class initialization.