Capture and use full path of Python executable
Previously, Python interpreter detection likely used "python",
"python2", "python3" as the Python interpreter path. This works
because command execution takes PATH into account.

But the determined Python path is ambiguous because it relies on
environment variables.

This commit changes the invoked Python script to print the value
of sys.executable. This value is captured and is used as the
Python interpreter path.