Expose _frozen and PyImport_FrozenModules
The _frozen struct and the PyImportFrozenModules variable are necessary
to enable a Rust program embedding Python to define custom frozen
modules.

I'm not a fan of the "_frozen" name for the struct. I'm not sure why it
doesn't have a proper name. But it is a documented struct
(https://docs.python.org/3/c-api/import.html#c._frozen).