Add "no-auto-initialize" feature to prevent call to prepare_freethreaded_python()
When embedding a Python interpreter in a Rust program, the automatic call
to prepare_freethreaded_python() to initialize the Python interpreter was
causing me problems.

Specifically, I'm doing very dirty things where I've managed to modify the
code that runs as part of Python interpreter startup, before Py_Initialize()
completes. The code that runs is using this crate. That code results
in a call to PyObject::drop(), which calls Python::acquire_gil(),
which calls GILGuard::acquire(), which calls prepare_freethreaded_python(),
which calls Py_InitializeEx(), and leads to reentrant execution because
Py_IsInitialized() returns 0 since the initial call hasn't completed.
I've managed to trigger a Rust panic due to detected reentrancy. I've
also managed to trigger deadlock, apparently due to the 2nd Py_Initialize()
trying to acquire the GIL, which is held by the original Py_Initialize().

Taking a step back, prepare_freethreaded_python() is making assumptions about
how to initialize the Python interpreter, which may vary from what applications
embedding Python may want. At the end of the day, the automatic call to this
function is to prevent a footgun where a consumer of this crate may forget
to initialize the Python interpreter first.

This commit adds a disabled-by-default (for backwards compatibility) feature
to disable the call to prepare_freethreaded_python() and therefore disables
the automatic initialization of the Python interpreter. When enabling the
feature, my Rust code using this crate that runs as part of Py_Initialize()
no longer results in reentrancy or GIL deadlocks.