py_capsule_fn! macro for direct function capsules
The `py_capsule_fn!` macro will make it easier (and safe) to
retrieve directly a function from a capsule.

Indeed the function type has to be specified on the fly, and
it wouldn't be safe to `transmute` without being sure we are
working on a function type. The retrieved function itself is
unsafe, like any function obtained through FFI should be.

The macro also adds a caching layer, using a static mut at the
top of the module it defines