Basic implementation for Python sets
This exposes the functions that are declared in ffi::setobject.
Since that latter doesn't declare iterators nor update (perhaps for
the same consistency reasons as for dicts ?), we don't have iteration
on Python sets (hence no listing as Vec), nor `update()`.

But this is already enough for Rust extensions to create Python sets
(our primary use-case) and to check membership.