Insisting on `static` for capsule creation from Rust
Creating a capsule pointing to some `const` definitely does
not work: that's what `static` is meant for.

Most of the creation methods had been written with mutable raw
pointers, leading to cast incompatibilities for actual `static`
data. They could have been worked around, but it's better to
insist on immutable pointers and refs from the ground up.

Chased all unnecessary occurrences of `*mut` and `&mut` away
in the process.