Capsules: access to underlying FFI PyObject
In case the encapsulated functions or structs make use of
C PyObject pointers, we need the corresponding Rust type to feed
methods that will convert into the high level PyObject.

This solution avoids to officially expose `ffi::PyObject` as part of the main
`rust-cpython` API (currently hidden in `cpython::_detail`).