Define inittab entities
The struct and function are documented at
https://docs.python.org/3/c-api/import.html#c._inittab
and defined in source code (in CPython master) at
https://github.com/python/cpython/blob/51edf8aaa2e17626f9690ed29d25945fc03016b9/Include/cpython/import.h#L30.

It is unclear whether PyImport_Inittab is part of the official
API, as it isn't clearly documented. But the static has been
shipping for years and is still in CPython master. I need to
use this variable in PyOxidizer in order to modify the set
of extensions at run-time. Putting it behind Py_LIMITED_API
should be safe.