implement property support
Add support for defining Python properties (attributes with getters and
optional setters) in Rust.

Properties are defined inside `py_class!` with:

    @property def property_name(&self) -> PyResult<PropertyType> { ... }

where `property_name` is the name in Python, and `PropertyType` is any type
that impls `ToPyObject`.

Property setters can optionally be defined as:

    @property_name.setter def set_property_name(
        &self,
        value: Option<PropertyType>
    ) -> PyResult<()> { ... }

If the value is `None` then the caller is deleting the property
(`del obj.property_name`).

If the value is `Some(x)` then the caller is setting the property to `x`.

For setters, the `PropertyType` can be any type that impls `FromPyObject`,
or a reference or optional reference to any type that impls `RefFromPyObject`.

In the latter case, the type of `value` is `Option<Option<&PropertyType>>`,
where `None` means the value is being deleted, `Some(None)` means the property
is being set to Python `None`, and `Some(Some(x))` means the property is being
set to `x`.