Make fields of initialization config public and implement Default
The structs added in #211 did not have public fields. The
fields need to be public in order to allow modifying them.

Furthermore, instantiating new instances of these structs would
be cumbersome (even with public fields) because there are so
many fields. So, we implement Default to construct new instances
that are zero initialized.