prevent unused_braces warnings in py_argparse_extract
The `py_argparse_extract` macro, for functions with no arguments,
can expand to expressions with unused braces.  This triggers
warnings like

    unnecessary braces around block return value

The block braces are necessary because we are passing the body through
declarative macros as a block, but once they are expanded to:

    Ok(()) => {
        let mut _iter = output.iter();
        {
            body
        }
    }

the compiler may consider the braces around the body to be superfluous
and raise a warning.

Work around this by assigning the value to a variable, which doesn't
trigger the warning.