Add a way to disable converting PyString to unicode on Python 2
The current logic that converts a PyString to either bytes or unicode
on Python 2 makes the `PyString` or `String` type less useful in
applications that pay closer attention to the actual type of bytes vs
string.

Add a feature, once set, PyString and String will be consistently
converted to the `str` type on both Python 2 and Python 3, regardless
of their actual content. This meets expections of type-sensitive
applications that still use Python 2.

The feature is disabled by default conservatively for backwards
compatibility.