Include Rust panic messages in Python exception
When Python calls a Rust function, an unhandled Rust panic is turned into
a Python `SystemError` exception. Previously, the value of that exception
was just "Rust panic" without more details.

This adds the panic message to that value.

Normally this message is printed to stderr by the Rust standard library’s
panic hook. In case that is not available for some reason, it can help
to duplicate it in the exception.