Add GILGuard::check() for python3
Adds a function that returns true if the GIL is currently held for the
current thread. PyGILState_Check() was only added in Python 3.4, so this
doesn't support Python 2.7.

When working with py.allow_threads(...) it is useful for the inner code
to be able to verify the GIL is not held in case it needs to acquire the
GIL on other threads. Being able to check this can help prevent
accidental deadlocks.

Ideally this could be done via static typing, similar to how the 'py'
object ensures the GIL is held, but I couldn't think of a clever way to
achieve this.